requirements.txt
boto3==1.14.28
botocore==1.17.29
PyYaml==5.3.1
fire==0.3.1
awscli==1.18.106
urllib3==1.22
requests==2.24.0